fix(observatory): secure production dependency graph
