+++ packages/phlo-observatory/src/phlo_observatory/package.json
@@ -24,45 +24,45 @@
     "@fontsource-variable/jetbrains-mono": "^5.2.8",
     "@primer/primitives": "^11.8.1",
     "@primer/react": "^38.22.0",
-    "@tailwindcss/vite": "^4.0.6",
     "@tanstack/react-query": "^5.101.0",
     "@tanstack/react-router": "^1.132.0",
-    "@tanstack/react-start": "^1.132.0",
     "@tanstack/react-table": "^8.21.3",
     "@tanstack/react-virtual": "^3.13.12",
-    "@tanstack/router-plugin": "^1.132.0",
     "@xyflow/react": "^12.10.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "lucide-react": "^1.8.0",
-    "nitro": "latest",
     "pino": "^10.1.0",
     "prism-react-renderer": "^2.4.1",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
-    "shadcn": "^3.6.1",
     "tailwind-merge": "^3.4.0",
-    "tailwindcss": "^4.0.6",
     "tw-animate-css": "^1.4.0",
-    "vite-tsconfig-paths": "^6.1.1",
-    "yaml": "^2.8.2",
+    "yaml": "^2.9.0",
     "zod": "^4.2.0"
   },
   "devDependencies": {
-    "@tanstack/devtools-vite": "^0.4.0",
+    "@tailwindcss/vite": "4.1.18",
+    "@tanstack/devtools-vite": "0.8.1",
     "@tanstack/eslint-config": "^0.3.0",
+    "@tanstack/react-start": "1.168.32",
+    "@tanstack/router-plugin": "1.168.23",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/react": "^16.2.0",
     "@types/node": "^25.5.2",
     "@types/react": "^19.2.0",
     "@types/react-dom": "^19.2.0",
     "@vitejs/plugin-react": "^5.0.4",
-    "esbuild": "^0.28.0",
+    "esbuild": "0.28.1",
     "jsdom": "^27.0.0",
+    "nitro": "3.0.260610-beta",
     "prettier": "^3.5.3",
+    "shadcn": "4.13.1",
+    "tailwindcss": "4.1.18",
     "typescript": "^5.7.2",
-    "vite": "^7.1.7",
+    "vite": "7.3.6",
+    "vite-tsconfig-paths": "^6.1.1",
     "vitest": "^4.1.4"
   }
 }
