+++ packages/phlo-observatory/src/phlo_observatory/src/components/GraphCanvas.tsx
@@ -22,7 +22,7 @@ import {
 import '@xyflow/react/dist/style.css'
 import { Database } from 'lucide-react'
 
-import type { GraphEdge, GraphNode } from '@/server/graph.server'
+import type { GraphEdge, GraphNode } from '@/observatory/api/graph'
 import type { Edge, Node, NodeProps, NodeTypes } from '@xyflow/react'
 
 const LAYER_STYLES: Record<
@@ -313,20 +313,18 @@ export function GraphCanvas({
   }, [graphNodes, focusedAsset, onAssetSelect])
 
   const initialEdges = useMemo(() => {
-    return graphEdges.map(
-      (edge, i): Edge => ({
-        id: `edge-${i}`,
-        source: edge.source,
-        target: edge.target,
-        type: 'smoothstep',
-        animated: false,
-        style: { stroke: 'var(--border)', strokeWidth: 2 },
-        markerEnd: {
-          type: MarkerType.ArrowClosed,
-          color: 'var(--border)',
-        },
-      }),
-    )
+    return graphEdges.map((edge, i): Edge => ({
+      id: `edge-${i}`,
+      source: edge.source,
+      target: edge.target,
+      type: 'smoothstep',
+      animated: false,
+      style: { stroke: 'var(--border)', strokeWidth: 2 },
+      markerEnd: {
+        type: MarkerType.ArrowClosed,
+        color: 'var(--border)',
+      },
+    }))
   }, [graphEdges])
 
   const [nodes, setNodes, onNodesChange] = useNodesState(initialNodes)
