+++ packages/phlo-observatory/src/phlo_observatory/src/extensions/registry.tsx
@@ -14,12 +14,12 @@ import type {
   ObservatoryExtension,
   ObservatoryExtensionNavItem,
   ObservatoryExtensionRoute,
-} from '@/server/extensions.server'
-import { getObservatoryExtensions } from '@/server/extensions.server'
+} from '@/observatory/api/extensions'
+import { getObservatoryExtensions } from '@/observatory/api/extensions'
 import {
   getExtensionSettings,
   putExtensionSettings,
-} from '@/server/extension-settings.server'
+} from '@/observatory/api/extension-settings'
 
 type ExtensionRouteContext = {
   createRoute: typeof createRoute
@@ -191,8 +191,7 @@ async function loadBrowserExtensions(): Promise<Array<ObservatoryExtension>> {
 
 function loadExtensionModule(moduleUrl: string): Promise<ExtensionModule> {
   const loader = bundledExtensionModules[moduleUrl] as
-    | ExtensionModuleLoader
-    | undefined
+    ExtensionModuleLoader | undefined
   if (loader) return loader()
   return Promise.resolve({})
 }
@@ -273,8 +272,7 @@ export function ObservatoryExtensionProvider({
               try {
                 const module = await loadExtensionModule(route.module)
                 const registerRoutes = module[route.export] as
-                  | RegisterRoutesFn
-                  | undefined
+                  RegisterRoutesFn | undefined
                 if (typeof registerRoutes !== 'function') return
                 const result = registerRoutes({
                   createRoute,
@@ -306,8 +304,7 @@ export function ObservatoryExtensionProvider({
               try {
                 const module = await loadExtensionModule(slot.module)
                 const registerSlotFn = module[slot.export] as
-                  | RegisterSlotFn
-                  | undefined
+                  RegisterSlotFn | undefined
                 if (typeof registerSlotFn !== 'function') return
                 registerSlotFn({ register: registerSlot(slot.slot_id) })
               } catch {
@@ -324,8 +321,7 @@ export function ObservatoryExtensionProvider({
               try {
                 const module = await loadExtensionModule(setting.module)
                 const registerSettingsFn = module[setting.export] as
-                  | RegisterSettingsFn
-                  | undefined
+                  RegisterSettingsFn | undefined
                 if (typeof registerSettingsFn !== 'function') return
                 const scope = extension.manifest.settings?.scope ?? 'extension'
                 const loadSettings = async () => {
