+++ packages/phlo-observatory/src/phlo_observatory/src/observatory/api/diff.ts
@@ -8,7 +8,7 @@
 import { createServerFn } from '@tanstack/react-start'
 
 import type { ColumnMapping, TransformType } from '@/utils/sqlParser'
-import { authMiddleware } from '@/server/auth.server'
+import { authMiddleware } from '@/observatory/api/auth'
 
 import { analyzeSQLTransformation } from '@/utils/sqlParser'
 
@@ -49,11 +49,7 @@ function computeDiffSummary(columnDiffs: Array<ColumnDiff>) {
  * Column change type for diff display
  */
 type ColumnChangeType =
-  | 'added'
-  | 'removed'
-  | 'renamed'
-  | 'transformed'
-  | 'unchanged'
+  'added' | 'removed' | 'renamed' | 'transformed' | 'unchanged'
 
 /**
  * Column diff entry
